SPIR-V

+++ Hazelnut/assets/shaders/Texture.glsl
@@ -1,4 +1,6 @@
-#Type vertex
+// Basic Texture Shader
+
+#type vertex
 #version 450 core
 
 layout(location = 0) in vec3 a_Position;
@@ -6,7 +8,7 @@ layout(location = 1) in vec4 a_Color;
 layout(location = 2) in vec2 a_TexCoord;
 layout(location = 3) in float a_TexIndex;
 layout(location = 4) in float a_TilingFactor;
-layout(location = 5) in float a_EntityID;
+layout(location = 5) in int a_EntityID;
 
 layout(std140, binding = 0) uniform Camera
 {
@@ -35,8 +37,7 @@ void main()
 	gl_Position = u_ViewProjection * vec4(a_Position, 1.0);
 }
 
-
-#Type fragment
+#type fragment
 #version 450 core
 
 layout(location = 0) out vec4 color;
@@ -94,7 +95,6 @@ void main()
 		case 30: texColor *= texture(u_Textures[30], Input.TexCoord * Input.TilingFactor); break;
 		case 31: texColor *= texture(u_Textures[31], Input.TexCoord * Input.TilingFactor); break;
 	}
-
 	color = texColor;
 
 	color2 = v_EntityID;
